namespace reorg + cleanup
I made sure that the namespaces I'm using for the game ECS stuff doesn't
conflict with the ones in the Friflo library. Also, I cleaned up the
Eyes component and the systems that use it. I'm planning to implement a
kind of memory for the eyes component.

Finally, this is my first commit description! :)